refactor(client): move stateStore.ts to its parent folder
