+++ packages/client/src/events/speaker.ts
@@ -1,5 +1,5 @@
 import { Dispatcher } from '../rtc/Dispatcher';
-import { StreamVideoWriteableStateStore } from '../stateStore';
+import { StreamVideoWriteableStateStore } from '../store';
 import { StreamVideoParticipantPatches } from '../rtc/types';
 
 /**
