+++ packages/client/src/store/index.ts
@@ -1,2 +1,2 @@
 export * as RxUtils from './rxUtils';
-export * from '../stateStore';
+export * from './stateStore';
